* Add 'expander-count' and 'expander-phys' for expander * Add SCSI enclosure service devices, right now just an empty device, no   any SCSI commands are handled.   TODO: add TEST UNIT READY, INQUIRY, RECEIVE DIAGNOSTIC, SEND   DIAGNOSTIC, READ BUFFER and WRITE BUFFER commands.

+++ hw/scsi/mpt3sas.h
@@ -183,10 +183,10 @@ struct MPT3SASState {
     MPT3SASEventQueue *event_queue;
 
     struct {
-        uint8_t count;
-        uint8_t all_phys;
-        uint8_t upstream_phys;
-        uint8_t downstream_phys;
+        uint32_t count;
+        uint32_t all_phys;
+        uint32_t upstream_phys;
+        uint32_t downstream_phys;
 
     } expander;
